feat: move widget to the module
